04.09.2023 Refactored

+++ src/main/java/telran/solution/dto/accounting/StatsDto.java
@@ -13,5 +13,5 @@ public class StatsDto {
     protected Integer checkedSolutions;
     @Setter
     protected Integer formulatedProblems;
-    protected Integer rating;
+    protected Double rating;
 }
